refactor tests to separated files
