+++ package.json
@@ -60,6 +60,7 @@
     "babel-preset-stage-2": "^6.22.0",
     "babel-preset-vue": "^2.0.1",
     "chalk": "^2.0.1",
+    "compression-webpack-plugin": "^1.1.11",
     "copy-webpack-plugin": "^4.0.1",
     "coveralls": "^3.0.0",
     "css-loader": "^0.28.10",
