+++ test/unit/specs/props.spec.js
@@ -1,7 +1,6 @@
 import Vue from 'vue'
 import { mount } from '@vue/test-utils'
 import rgb from 'rgb'
-import isHexColor from 'is-hex-color'
 
 import Swatches, {
   DEFAULT_BACKGROUND_COLOR,
@@ -10,7 +9,7 @@ import Swatches, {
   DEFAULT_SWATCH_SIZE
 } from 'src/Swatches'
 import Swatch from 'src/Swatch'
-import presets, { supportedProperties } from 'src/presets'
+import presets from 'src/presets'
 
 const completPresetExample = {
   swatches: ['#cc4125', '#e06666', '#f6b26b', '#ffd966', '#93c47d', '#76a5af', '#6d9eeb', '#6fa8dc', '#8e7cc3', '#c27ba0'],
@@ -1160,388 +1159,5 @@ describe('Props', () => {
         .toEqual(0)
       })
     })
-  })
-})
-
-describe('Events', () => {
-  describe('@input', () => {
-    test('should be emited whenever user pick a swatch', () => {
-      const componentWrapper = mount(Swatches)
-      const swatch = componentWrapper.find(Swatch)
-      swatch.trigger('click')
-
-      return Vue.nextTick()
-      .then(() => {
-        expect(componentWrapper.emitted().input.length).toEqual(1)
-      })
-    })
-    test('should payload the value', () => {
-      const componentWrapper = mount(Swatches)
-      const swatch = componentWrapper.find(Swatch)
-      swatch.trigger('click')
-      const color = swatch.vm.swatchColor
-
-      return Vue.nextTick()
-      .then(() => {
-        expect(componentWrapper.emitted().input[0][0])
-        .toEqual(color)
-      })
-    })
-  })
-  describe('@open', () => {
-    test('should not be emited when Inline mode is activated', () => {
-      const componentWrapper = mount(Swatches, {
-        propsData: {
-          inline: true
-        }
-      })
-      componentWrapper.vm.showPopover()
-
-      return Vue.nextTick()
-      .then(() => {
-        expect(componentWrapper.emitted().open)
-        .not.toBeTruthy()
-      })
-    })
-    test('should be emited when Inline mode is not activated', () => {
-      const componentWrapper = mount(Swatches, {
-        propsData: {
-          inline: false
-        }
-      })
-      componentWrapper.vm.showPopover()
-
-      return Vue.nextTick()
-      .then(() => {
-        expect(componentWrapper.emitted().open.length).toEqual(1)
-      })
-    })
-    describe('When default trigger is used', () => {
-      test('should be emited whenever user clicks the trigger', () => {
-        const componentWrapper = mount(Swatches, {
-          propsData: {
-            inline: false
-          }
-        })
-        const trigger = componentWrapper.find({ ref: 'trigger-wrapper' })
-        trigger.trigger('click')
-
-        return Vue.nextTick()
-        .then(() => {
-          expect(componentWrapper.emitted().open.length).toEqual(1)
-        })
-      })
-    })
-    describe('When trigger slot is used', () => {
-      test('should be emited whenever user clicks the trigger', () => {
-        const buttonTest = '<button id="button-test">Hello World</button>'
-        const componentWrapper = mount(Swatches, {
-          slots: {
-            trigger: buttonTest
-          },
-          propsData: {
-            inline: false
-          }
-        })
-        const trigger = componentWrapper.find('#button-test')
-        trigger.trigger('click')
-
-        return Vue.nextTick()
-        .then(() => {
-          expect(componentWrapper.emitted().open.length).toEqual(1)
-        })
-      })
-    })
-  })
-  describe('@close', () => {
-    test('should not be emited when Inline mode is activated', () => {
-      const componentWrapper = mount(Swatches, {
-        propsData: {
-          inline: true,
-          closeOnSelect: true
-        }
-      })
-      componentWrapper.vm.onBlur(null)
-
-      return Vue.nextTick()
-      .then(() => {
-        expect(componentWrapper.emitted().close)
-        .not.toBeTruthy()
-      })
-    })
-    describe('When default trigger is used', () => {
-      test('should be emited whenever user close the Popover by clicking outside', () => {
-        const componentWrapper = mount(Swatches, {
-          propsData: {
-            inline: false
-          }
-        })
-        componentWrapper.vm.showPopover()
-        componentWrapper.vm.onBlur(null)
-
-        return Vue.nextTick()
-        .then(() => {
-          expect(componentWrapper.emitted().close.length).toEqual(1)
-        })
-      })
-      test('should be emited whenever user close the Popover by clicking the trigger', () => {
-        const componentWrapper = mount(Swatches, {
-          propsData: {
-            inline: false
-          }
-        })
-        const trigger = componentWrapper.find({ ref: 'trigger-wrapper' })
-        trigger.trigger('click') // one click to open the Popover
-        trigger.trigger('click') // and another to close the Popover
-
-        return Vue.nextTick()
-        .then(() => {
-          expect(componentWrapper.emitted().close.length).toEqual(1)
-        })
-      })
-      test('should be emited whenever user close the Popover by clicking a Swatch', () => {
-        const componentWrapper = mount(Swatches, {
-          propsData: {
-            inline: false,
-            closeOnSelect: true
-          }
-        })
-        componentWrapper.vm.showPopover()
-        const swatch = componentWrapper.find(Swatch)
-        swatch.trigger('click')
-
-        return Vue.nextTick()
-        .then(() => {
-          expect(componentWrapper.emitted().close.length).toEqual(1)
-        })
-      })
-    })
-    describe('When trigger slot is used', () => {
-      test('should be emited whenever user close the Popover by clicking outside', () => {
-        const buttonTest = '<button id="button-test">Hello World</button>'
-        const componentWrapper = mount(Swatches, {
-          slots: {
-            trigger: buttonTest
-          },
-          propsData: {
-            inline: false
-          }
-        })
-        componentWrapper.vm.showPopover()
-        componentWrapper.vm.onBlur(null)
-
-        return Vue.nextTick()
-        .then(() => {
-          expect(componentWrapper.emitted().close.length).toEqual(1)
-        })
-      })
-      test('should be emited whenever user close the Popover by clicking the trigger', () => {
-        const buttonTest = '<button id="button-test">Hello World</button>'
-        const componentWrapper = mount(Swatches, {
-          slots: {
-            trigger: buttonTest
-          },
-          propsData: {
-            inline: false
-          }
-        })
-        const trigger = componentWrapper.find({ ref: 'trigger-wrapper' })
-        trigger.trigger('click')
-        trigger.trigger('click')
-
-        return Vue.nextTick()
-        .then(() => {
-          expect(componentWrapper.emitted().close.length).toEqual(1)
-        })
-      })
-      test('should be emited whenever user close the Popover by clicking a Swatch', () => {
-        const buttonTest = '<button id="button-test">Hello World</button>'
-        const componentWrapper = mount(Swatches, {
-          slots: {
-            trigger: buttonTest
-          },
-          propsData: {
-            inline: false
-          }
-        })
-        componentWrapper.vm.showPopover()
-        const swatch = componentWrapper.find(Swatch)
-        swatch.trigger('click')
-
-        return Vue.nextTick()
-        .then(() => {
-          expect(componentWrapper.emitted().close.length).toEqual(1)
-        })
-      })
-    })
-    test('should payload the value', () => {
-      const componentWrapper = mount(Swatches, {
-        propsData: {
-          inline: false,
-          closeOnSelect: true
-        }
-      })
-      componentWrapper.vm.showPopover()
-      const swatch = componentWrapper.find(Swatch)
-      swatch.trigger('click')
-      const color = swatch.vm.swatchColor
-
-      return Vue.nextTick()
-      .then(() => {
-        expect(componentWrapper.emitted().close[0][0])
-        .toEqual(color)
-      })
-    })
-  })
-})
-
-describe('Slots', () => {
-  describe('trigger', () => {
-    test('should replace the trigger node', () => {
-      const spanTest = '<span id="span-test">Hello World</span>'
-      const componentWrapper = mount(Swatches, {
-        slots: {
-          trigger: spanTest
-        },
-        propsData: {
-          inline: false
-        }
-      })
-      const trigger = componentWrapper.find('#span-test')
-
-      return Vue.nextTick()
-      .then(() => {
-        expect(trigger.html())
-        .toEqual(spanTest)
-      })
-    })
-  })
-})
-
-describe('Presets', () => {
-  let presetsTested = 0
-  const testPreset = (presetName) => {
-    describe(`"${presetName}"`, () => {
-      test('should have swatches property', () => {
-        const preset = presets[presetName]
-        const swatches = preset.swatches
-
-        expect(swatches)
-        .toBeTruthy()
-      })
-      test('preset should have valid colors', () => {
-        let isNested = false
-        let validSwatches = 0
-        let swatchesCount = 0
-        const preset = presets[presetName]
-        const swatches = preset.swatches
-
-        if (swatches instanceof Array && swatches.length > 0 && swatches[0] instanceof Array) {
-          isNested = true
-        }
-        if (isNested) {
-          swatches.forEach(r => {
-            if (!(r instanceof Array)) return swatchesCount++
-            r.forEach(s => {
-              if (isHexColor(s)) validSwatches++
-              return swatchesCount++
-            })
-          })
-        } else {
-          swatches.forEach(s => {
-            if (isHexColor(s)) validSwatches++
-            return swatchesCount++
-          })
-        }
-
-        expect(`${validSwatches} valid colors`)
-        .toEqual(`${swatchesCount} valid colors`)
-      })
-      test('preset should have a valid borderRadius', () => {
-        let valid = false
-        const preset = presets[presetName]
-        const borderRadius = preset.borderRadius
-        if (typeof borderRadius === 'undefined') valid = true // borderRadius is not required
-        if (typeof borderRadius === 'string') valid = true
-
-        expect(valid)
-        .toBeTruthy()
-      })
-      test('preset should have a valid rowLength', () => {
-        let valid = false
-        const preset = presets[presetName]
-        const rowLength = preset.rowLength
-        if (typeof rowLength === 'undefined') valid = true // rowLength is not required
-        if (typeof rowLength === 'number' && rowLength > 0 && Number.isInteger(rowLength)) valid = true
-
-        expect(valid)
-        .toBeTruthy()
-      })
-      test('preset should have a valid spacingSize', () => {
-        let valid = false
-        const preset = presets[presetName]
-        const spacingSize = preset.spacingSize
-        if (typeof spacingSize === 'undefined') valid = true // spacingSize is not required
-        if (typeof spacingSize === 'number' && spacingSize >= 0) valid = true
-
-        expect(valid)
-        .toBeTruthy()
-      })
-      test('preset should have a valid swatchSize', () => {
-        let valid = false
-        const preset = presets[presetName]
-        const swatchSize = preset.swatchSize
-        if (typeof swatchSize === 'undefined') valid = true // swatchSize is not required
-        if (typeof swatchSize === 'number' && swatchSize > 0) valid = true
-
-        expect(valid)
-        .toBeTruthy()
-      })
-      test('preset should have a valid maxHeight', () => {
-        let valid = false
-        const preset = presets[presetName]
-        const maxHeight = preset.maxHeight
-        if (typeof maxHeight === 'undefined') valid = true // maxHeight is not required
-        if (typeof maxHeight === 'number' && maxHeight > 0) valid = true
-
-        expect(valid)
-        .toBeTruthy()
-      })
-      test('preset should have a valid showBorder', () => {
-        let valid = false
-        const preset = presets[presetName]
-        const showBorder = preset.showBorder
-        if (typeof showBorder === 'undefined') valid = true // showBorder is not required
-        if (typeof showBorder === 'boolean') valid = true
-
-        expect(valid)
-        .toBeTruthy()
-      })
-      test('preset should only have a supported properties', () => {
-        let validProperties = 0
-        const preset = presets[presetName]
-        const keys = Object.keys(preset)
-
-        keys.forEach(k => {
-          if (supportedProperties.indexOf(k) !== -1) validProperties++
-        })
-
-        expect(`${keys.length} valid properties`)
-        .toEqual(`${validProperties} valid properties`)
-      })
-      presetsTested++
-    })
-  }
-
-  testPreset('basic')
-  testPreset('material-basic')
-  testPreset('material-dark')
-  testPreset('material-light')
-  testPreset('text-basic')
-  testPreset('text-advanced')
-
-  test('all presets should be tested', () => {
-    expect(presetsTested)
-    .toEqual(Object.keys(presets).length)
   })
 })
